Split parsers out of zstd_compress.c
